requirements.txt
python-slugify==4.0.0
hdx-python-api==4.3.5
pygsheets==2.0.3.1